requirements.txt
aiogram>=3.0.0
aiosqlite
feedparser
httpx
python-dateutil
